Cache.py now able to be used with main.py

+++ main.py
@@ -1,13 +1,12 @@
 import discord
 from discord.ext import commands
-import asyncio
-from concurrent.futures import ThreadPoolExecutor
+
 import logging
+
 import jstrisfunctions
 from jstrisfunctions import DateInit
 from jstrisfunctions import IndivParameterInit
-from jstrisuser import UserIndivGames
-from jstrisuser import UserLiveGames
+from cache import CacheInit
 
 
 intents = discord.Intents.default()
@@ -15,7 +14,6 @@ intents.members = True
 description = 'Third party BadgerBot to quickly gather Jstris stats on individual players'
 BadgerBot = commands.Bot(command_prefix='?', description=description, intents=intents, help_command=None)
 
-LOOP = asyncio.get_event_loop()
 num_processes = 0
 
 logging.basicConfig(level=logging.INFO, filename="logjstris.log", datefmt='%m/%d/%Y %H:%M:%S',
@@ -75,9 +73,8 @@ class IndivCommands(commands.Cog):
         my_ps = IndivParameterInit(args)
 
         init_message = await ctx.send(f"Searching {username}'s games now. This can take a while.")
-        searched_games = await LOOP.run_in_executor(ThreadPoolExecutor(),
-                                                    UserIndivGames,
-                                                    username, my_ps.game, my_ps.mode, my_ps.first_date, my_ps.last_date)
+        searched_games = CacheInit(username, my_ps)
+        await searched_games.fetch_all_games()
         await GeneralMaintenance.num_processes_finish()
 
         await init_message.delete()
@@ -85,7 +82,7 @@ class IndivCommands(commands.Cog):
             await ctx.send(ctx.author.mention)
             await ctx.send(searched_games.error_message)
         else:
-            a = jstrisfunctions.least_(searched_games.all_replays, my_ps.param)
+            a = jstrisfunctions.least_(searched_games.in_period_replays, my_ps.param)
             await IndivCommands.replay_send(ctx, a)
 
         logging.info("Finish least")
@@ -100,9 +97,9 @@ class IndivCommands(commands.Cog):
         my_ps = IndivParameterInit(args)
 
         init_message = await ctx.send(f"Searching {username}'s games now. This can take a while.")
-        searched_games = await LOOP.run_in_executor(ThreadPoolExecutor(),
-                                                    UserIndivGames,
-                                                    username, my_ps.game, my_ps.mode, my_ps.first_date, my_ps.last_date)
+        searched_games = CacheInit(username, my_ps)
+        await searched_games.fetch_all_games()
+
         await GeneralMaintenance.num_processes_finish()
 
         await init_message.delete()
@@ -110,7 +107,7 @@ class IndivCommands(commands.Cog):
             await ctx.send(ctx.author.mention)
             await ctx.send(searched_games.error_message)
         else:
-            a = jstrisfunctions.most_(searched_games.all_replays, my_ps.param)
+            a = jstrisfunctions.most_(searched_games.in_period_replays, my_ps.param)
             await IndivCommands.replay_send(ctx, a)
         logging.info("Finishing most")
 
@@ -123,16 +120,16 @@ class IndivCommands(commands.Cog):
         my_ps = IndivParameterInit(args)
 
         init_message = await ctx.send(f"Searching {username}'s games now. This can take a while.")
-        searched_games = await LOOP.run_in_executor(ThreadPoolExecutor(),
-                                                    UserIndivGames,
-                                                    username, my_ps.game, my_ps.mode, my_ps.first_date, my_ps.last_date)
+        searched_games = CacheInit(username, my_ps)
+        await searched_games.fetch_all_games()
+
         await GeneralMaintenance.num_processes_finish()
         await init_message.delete()
         if searched_games.has_error:
             await ctx.send(ctx.author.mention)
             await ctx.send(searched_games.error_message)
         else:
-            my_average = jstrisfunctions.average_(searched_games.all_replays, my_ps.param)
+            my_average = jstrisfunctions.average_(searched_games.in_period_replays, my_ps.param)
             await ctx.send(f"Average {my_ps.param} for {username} is: {my_average}")
         logging.info("Finishing average")
 
@@ -144,9 +141,9 @@ class IndivCommands(commands.Cog):
 
         my_ps = IndivParameterInit(args)
         init_message = await ctx.send(f"Searching {username}'s games now. This can take a while.")
-        searched_games = await LOOP.run_in_executor(ThreadPoolExecutor(),
-                                                    UserIndivGames,
-                                                    username, my_ps.game, my_ps.mode, my_ps.first_date, my_ps.last_date)
+        searched_games = CacheInit(username, my_ps)
+        await searched_games.fetch_all_games()
+
         await GeneralMaintenance.num_processes_finish()
 
         await init_message.delete()
@@ -154,7 +151,7 @@ class IndivCommands(commands.Cog):
             await ctx.send(ctx.author.mention)
             await ctx.send(searched_games.error_message)
         else:
-            a = len(searched_games.all_replays)
+            a = len(searched_games.in_period_replays)
             await ctx.send(f"{str(a)} games")
         logging.info("Finishing numgames")
 
@@ -170,17 +167,21 @@ class IndivCommands(commands.Cog):
 
         init_message = await ctx.send(f"Searching {username}'s games now. This can take a while.")
 
+        args = list(args)
+        args.append('cheese')
+        args = tuple(args)
+
         my_ps = IndivParameterInit(args)
-        searched_games = await LOOP.run_in_executor(ThreadPoolExecutor(),
-                                                    UserIndivGames,
-                                                    username, "3", "3", my_ps.first_date, my_ps.last_date)
+        searched_games = CacheInit(username, my_ps)
+        await searched_games.fetch_all_games()
+
         await GeneralMaintenance.num_processes_finish()
         await init_message.delete()
         if searched_games.has_error:
             await ctx.send(ctx.author.mention)
             await ctx.send(searched_games.error_message)
         else:
-            a = jstrisfunctions.subblocks(searched_games.all_replays, int(blocks))
+            a = jstrisfunctions.subblocks(searched_games.in_period_replays, int(blocks))
             await ctx.send(ctx.author.mention)
             await ctx.send(f"{username} has {a} sub {blocks}s")
         logging.info("Finishing subblocks")
@@ -210,15 +211,19 @@ class VsCommands(commands.Cog):
         self.bot = bot
 
     @commands.command()
-    async def vs(self, ctx, username: str, offset: int = 10):
+    async def vs(self, ctx, username: str, offset: int = 10, first_date: str = "1000 months", last_date: str = "0 days"):
         logging.info("Beginning vs")
         if not await GeneralMaintenance.num_processes_init(ctx):
             return None
 
+        date_init = DateInit(first_date, last_date)
+        if date_init.has_error:
+            await ctx.send(date_init.error_message)
+            return 0
+
         init_message = await ctx.send(f"Searching {username}'s games now. This can take a while.")
-        searched_games = await LOOP.run_in_executor(ThreadPoolExecutor(),
-                                                    UserLiveGames,
-                                                    username, offset)
+        searched_games = CacheInit(username, date_init)
+
         await GeneralMaintenance.num_processes_finish()
         await init_message.delete()
         if searched_games.has_error:
@@ -227,17 +232,17 @@ class VsCommands(commands.Cog):
             return None
 
         # Calculates averages
-        apm_avg = jstrisfunctions.live_games_avg(searched_games.all_replays, offset, 'apm')
-        spm_avg = jstrisfunctions.live_games_avg(searched_games.all_replays, offset, 'spm')
-        pps_avg = jstrisfunctions.live_games_avg(searched_games.all_replays, offset, 'pps')
-        weight_apm = round(jstrisfunctions.live_games_weighted_avg(searched_games.all_replays, offset, 'attack') * 60,
+        apm_avg = jstrisfunctions.live_games_avg(searched_games.in_period_replays, offset, 'apm')
+        spm_avg = jstrisfunctions.live_games_avg(searched_games.in_period_replays, offset, 'spm')
+        pps_avg = jstrisfunctions.live_games_avg(searched_games.in_period_replays, offset, 'pps')
+        weight_apm = round(jstrisfunctions.live_games_weighted_avg(searched_games.in_period_replays, offset, 'attack') * 60,
                            2)
-        weight_spm = round(jstrisfunctions.live_games_weighted_avg(searched_games.all_replays, offset, 'sent') * 60, 2)
-        weight_pps = round(jstrisfunctions.live_games_weighted_avg(searched_games.all_replays, offset, 'pcs'), 2)
-        time_avg = jstrisfunctions.live_games_avg(searched_games.all_replays, offset, 'gametime')
-        players_avg = jstrisfunctions.live_games_avg(searched_games.all_replays, offset, 'players')
-        pos_avg = jstrisfunctions.live_games_avg(searched_games.all_replays, offset, 'pos')
-        won_games = jstrisfunctions.games_won(searched_games.all_replays, offset)
+        weight_spm = round(jstrisfunctions.live_games_weighted_avg(searched_games.in_period_replays, offset, 'sent') * 60, 2)
+        weight_pps = round(jstrisfunctions.live_games_weighted_avg(searched_games.in_period_replays, offset, 'pcs'), 2)
+        time_avg = jstrisfunctions.live_games_avg(searched_games.in_period_replays, offset, 'gametime')
+        players_avg = jstrisfunctions.live_games_avg(searched_games.in_period_replays, offset, 'players')
+        pos_avg = jstrisfunctions.live_games_avg(searched_games.in_period_replays, offset, 'pos')
+        won_games = jstrisfunctions.games_won(searched_games.in_period_replays, offset)
 
         # Discord formatting
         embed = await embed_init(username)
@@ -251,9 +256,9 @@ class VsCommands(commands.Cog):
         embed.add_field(name="**final position:**", value=str(pos_avg), inline=True)
         embed.add_field(name="**players:**", value=str(players_avg), inline=True)
         embed.add_field(name="**games won:**", value=f"{won_games}  "
-                                                     f"({won_games / len(searched_games.all_replays) * 100:.2f}%)",
+                                                     f"({won_games / len(searched_games.in_period_replays) * 100:.2f}%)",
                         inline=False)
-        embed.add_field(name="**number of games:**", value=str(len(searched_games.all_replays)), inline=False)
+        embed.add_field(name="**number of games:**", value=str(len(searched_games.in_period_replays)), inline=False)
         embed.set_footer(text='All of these values are averages. Weighted means weighted by time, not game.')
         await ctx.send(ctx.author.mention)
         await ctx.send(embed=embed)
@@ -267,23 +272,20 @@ class VsCommands(commands.Cog):
             await ctx.send(date_init.error_message)
             return 0
 
-        first_date = date_init.first
-        last_date = date_init.last
-
         if not await GeneralMaintenance.num_processes_init(ctx):
             return None
 
         init_message = await ctx.send(f"Searching {username}'s games now. This can take a while.")
-        searched_games = await LOOP.run_in_executor(ThreadPoolExecutor(),
-                                                    UserLiveGames,
-                                                    username, 1000000000, first_date, last_date)
+        searched_games = CacheInit(username, date_init)
+        await searched_games.fetch_all_games()
+
         await GeneralMaintenance.num_processes_finish()
         await init_message.delete()
         if searched_games.has_error:
             await ctx.send(ctx.author.mention)
             await ctx.send(searched_games.error_message)
             return None
-        list_of_opponents = jstrisfunctions.opponents_matchups(searched_games.all_replays)
+        list_of_opponents = jstrisfunctions.opponents_matchups(searched_games.in_period_replays)
 
         # Discord formatting stuff
 
@@ -326,39 +328,35 @@ class VsCommands(commands.Cog):
             await ctx.send(date_init.error_message)
             return 0
 
-        first_date = date_init.first
-        last_date = date_init.last
-
         if not await GeneralMaintenance.num_processes_init(ctx):
             return None
         init_message = await ctx.send(f"Searching {username}'s games now. This can take a while.")
 
         # Username's games
         logging.info(f"Beginning {username}")
-        searched_games = await LOOP.run_in_executor(ThreadPoolExecutor(),
-                                                    UserLiveGames,
-                                                    username, 10000000000, first_date, last_date)
+        searched_games = CacheInit(username, date_init)
+        await searched_games.fetch_all_games()
 
         if searched_games.has_error:
             await ctx.send(ctx.author.mention)
             await ctx.send(searched_games.error_message)
             return None
-        list_of_opponents = jstrisfunctions.opponents_matchups(searched_games.all_replays)
+        list_of_opponents = jstrisfunctions.opponents_matchups(searched_games.in_period_replays)
         embed1 = await VsCommands.vs_matchup_embed(ctx, username, opponent, list_of_opponents)
 
         # Opponent's games
         logging.info(f"Beginning {opponent}, first: {list_of_opponents[opponent]['min_time']}, "
                      f"last: {list_of_opponents[opponent]['max_time']}")
-        opp_first_date = list_of_opponents[opponent]["min_time"]
-        opp_last_date = list_of_opponents[opponent]["max_time"]
-        searched_games = await LOOP.run_in_executor(ThreadPoolExecutor(),
-                                                    UserLiveGames,
-                                                    opponent, 10000000, opp_first_date, opp_last_date)
+        date_init.first = list_of_opponents[opponent]["min_time"]
+        date_init.last = list_of_opponents[opponent]["max_time"]
+        searched_games = CacheInit(username, date_init)
+        await searched_games.fetch_all_games()
+
         if searched_games.has_error:
             await ctx.send(ctx.author.mention)
             await ctx.send(searched_games.error_message)
             return None
-        list_of_opponents = jstrisfunctions.opponents_matchups(searched_games.all_replays)
+        list_of_opponents = jstrisfunctions.opponents_matchups(searched_games.in_period_replays)
         embed2 = await VsCommands.vs_matchup_embed(ctx, opponent, username, list_of_opponents)
 
         # Finalizing
@@ -397,7 +395,7 @@ class VsCommands(commands.Cog):
     #         await ctx.send(searched_games.error_message)
     #         return None
     #
-    #     all_games = searched_games.all_replays
+    #     all_games = searched_games.all_replays_ever
     #     list_of_times = [datetime.datetime.strptime(game['gtime'], "%Y-%m-%d %H:%M:%S") for game in all_games]
     #     list_of_apms = [game['apm'] for game in all_games]
     #     list_of_sizes = [5 for game in list_of_times]
@@ -473,6 +471,8 @@ if __name__ == "__main__":
 
 
 # To do list
-# Make min and max time into a standard function
 # indiv stats for apm and such
 # Complete the cache
+
+# finish filter
+# finish pruning stats for storage and readding stats for presentation
